requirements.txt
# requirements.txt

flask==1.1.2
flask_sqlalchemy==2.4.4
flask_migrate==2.5.3
flask_login==0.5.0
flask_wtf==0.14.3
flask_bootstrap==3.3.7.1
python_dotenv==0.15.0
azure==5.0.0
azure-storage-blob==12.8.0
stripe==2.55.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-auth-httplib2==0.1.0
google-api-python-client==2.2.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-meet-python-api==0.1.1
googlemaps==4.4.5
pinecone-client==0.0.2
langchain==0.1.0
mistral-api==0.1.0
sqlalchemy==1.3.23
psycopg2-binary==2.8.6
gunicorn==20.0.4
pytest==6.2.2
coverage==5.5
flake8==3.8.4
black==20.8b1
isort==5.7.0
pre-commit==2.11.1
